Add a --debug option to provide details, requires click
If this is run without logging into Humble just prior, especially
if there are multiple browser profiles, it could pick the wrong
one.

Add a debug option to display the candidates and the chosen
path on stderr to give the user some feedback on progress.

Use the click library instead of sys.arg* for the cookies path name
if provided directly.

Signed-off-by: Rob Crittenden <rcrit@greyoak.com>